requirements.txt
torch
torchvision
kagglehub
Pillow
tqdm
huggingface_hub # For pushing to Hugging Face
numpy